Admin Dashboard Addedgit add .

+++ node-server/src/controllers/userController.js
@@ -1,4 +1,44 @@
 const User = require("../models/userModel");
+const path = require('path');
+
+
+// This API is used to create a new user with given user data
+// HANDLED BY MULTER SEE ROUTE FOR MORE DETAILS
+exports.createUser = async (req, res) => {
+    try {
+        const { email, username, firstname, lastname, password} = req.body;
+        
+        let userData = { email, username, firstname, lastname, password};
+        if(req.file){
+            const photo = req.file.filename;
+            userData = { ...userData,photo};
+        }
+        // console.log(req.file);
+        // console.log("org name",req.file.originalname);
+        // console.log("name",req.filename);
+        // console.log("photo:",photo);
+        const user = new User(userData);
+        // console.log(user);
+        await user.save();
+        const token = user.getJWTToken();
+
+        res.status(200).json({
+            success: true,
+            message: "User Created Successfully !!!",
+            token: token
+        });
+    } catch (err) {
+        res.status(err.statusCode || 400).json({ success: false, message: err.message });
+    }
+
+}
+
+// Send Profile Image
+exports.getImageByAuth = async(req,res) => {
+    res.sendFile( 'defaultProfile.png' , {root: "./uploads"});
+    // res.json({success: true});
+    
+}
 
 
 // Account START 
@@ -44,6 +84,9 @@ exports.updatePasswordByAuth = async(req,res) => {
 // Account END 
 
 // This API used to retrieve all the users from the database.
+
+
+// ADMIN PROTECTED ROUTES.
 exports.getAllUsers = async (req, res) => {
     try {
         const users = await User.find({});
@@ -70,32 +113,32 @@ exports.getUserByID = async (req, res) => {
 
 }
 
-// This API is used to create a new user with given user data
-exports.createUser = async (req, res) => {
-    try {
-        const userData = req.body;
-        const user = new User(userData);
-        console.log(user);
-        await user.save();
-        res.status(200).json({ success: true, message: "User Created Successfully !!!" });
-    } catch (err) {
-        res.status(err.statusCode || 400).json({ success: false, message: err.message });
-    }
-
-}
 
 
 // This API used to update details of a given user (identified by id).
 exports.updateUserByID = async (req, res) => {
     try {
         const id = req.params.id;
-        const {firstname,lastname,age} = req.body;
-        const userData = {first_name,last_name,age};
-        const user = await User.findByIdAndUpdate(id, userData);
+        // const {firstname,lastname,age} = req.body;
+        // const userData = {first_name,last_name,age};
+
+        if (id == req.user._id) {
+            res.status(405).json({ success: false, message: "You Can't Update you're own Account..." })
+            return;
+        }
+        let userData = req.body; 
+        console.log(req.body);
+        if (req.file) {
+            const photo = req.file.filename;
+            userData = { ...userData, photo };
+        }
+        // console.log(userData);
+        const user = await User.findByIdAndUpdate(id, userData,{new: true});
+        console.log(user);
         if (!user) {
             res.status(200).json({ success: false, message: "User with given id not found!" });
         } else {
-            res.status(200).json({ success: true, message: "User Updated Successfully" });
+            res.status(200).json({ success: true, message: "User Updated Successfully", data: user });
         }
     } catch (err) {
         res.status(err.statusCode || 400).json({ success: false, message: err.message });
@@ -106,9 +149,31 @@ exports.updateUserByID = async (req, res) => {
 }
 
 // This API used to delete the given user (identified by id)
+// Admin Route...
+// Send Profile Image   
+// For testing removing Admin route, and auth route.
+exports.getImageByID = async (req, res) => {
+    try{
+        const id = req.params.id;
+        const user = await User.findById(id);
+        res.sendFile(user.photo, { root: "./uploads" });
+        // res.sendFile('defaultProfile.png', { root: "./uploads" });
+    }catch(err){
+        res.json({success: false, message: err.message});
+    }
+  
+}
+
+// This API used to delete the given user (identified by id)
+// Admin Route...
 exports.deleteUserByID = async (req, res) => {
     try {
         const id = req.params.id;
+        // console.log(id,"\n\n", req.user, "\n\n", req.user._id);
+        if ( id == req.user._id) {
+            res.status(405).json({ success: false, message: "You Can't delete you're own Account..." })
+            return;
+        }
         const user = await User.findByIdAndDelete(id);
         if (!user) {
             res.status(200).json({ success: false, message: "User with given id not found!" });
@@ -120,6 +185,12 @@ exports.deleteUserByID = async (req, res) => {
     }
 }
 
+
+
+
+
+
+// LOGIN/SIGN IN  - LOGOUT/SIGN OUT FUNCTIONALITIES...
 // for showing the authentication functionalities
 
 exports.loginUser = async (req, res, next) => {
@@ -164,7 +235,7 @@ exports.loginUser = async (req, res, next) => {
         //     message: "logged in successfully",
         //     token: token
         // });
-        console.log(token);
+        // console.log(token);
         res.status(200).json({
             success: true,
             message: "logged in successfully",
